--- src/components/Contacts/Contacts.scss
@@ -1,36 +1,0 @@
-@import "../../styles/utils/variables";
-
-.Contacts-container {
-  max-width: 1072px;
-  margin: 0 auto;
-  padding: 0 16px;
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-
-  .Contacts-form {
-    display: flex;
-    flex-direction: column;
-    gap: 12px;
-
-    input,
-    textarea {
-      padding: 6px 12px;
-      background-color: rgba(236, 236, 236, 1);
-    }
-
-    textarea {
-      resize: vertical;
-      min-height: 80px;
-    }
-
-    .Contacts-btn {
-      background-color: rgba(18, 32, 45, 1);
-      color: #fff;
-      border-bottom: 2px solid rgba(220, 182, 90, 1);
-      padding: 7px 0;
-      font-family: $font-semibold;
-      font-size: 16px;
-    }
-  }
-}
