+++ src/components/OrderEasily/OrderEasily.scss
@@ -20,6 +20,12 @@
       padding: 6px 0px 6px 12px;
       background-color: rgba(236, 236, 236, 1);
     }
+    textarea {
+      padding: 6px 12px;
+      background-color: rgba(236, 236, 236, 1);
+      resize: vertical;
+      min-height: 80px;
+    }
     p {
       font-family: $font-light;
       font-weight: 300;
