+++ src/pages/Home/Home.jsx
@@ -5,7 +5,6 @@ import CardItem from "../../components/CardItem/CardItem";
 import ChoseProffesional from "../../components/ChoseProffesional/ChoseProffesional";
 import NewProducts from "../../components/NewProducts/NewProducts";
 import OrderEasily from "../../components/OrderEasily/OrderEasily";
-import Contacts from "../../components/Contacts/Contacts";
 import PopularBrands from "../../components/PopularBrands/PopularBrands";
 import ProductCategories from "../../components/ProductCategories/ProductCategories";
 import Reviews from "../../components/Reviews/Reviews";
@@ -25,7 +24,6 @@ function Home() {
       <Advantages />
       <Reviews />
       <OrderEasily />
-      <Contacts />
     </>
   );
 }
